requirements.txt
openai==1.13.3
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydub==0.25.1
jinja2==3.1.3
python-multipart==0.0.9
httpx==0.27.0
httpcore==1.0.4
python-jose==3.3.0
